P2P class to create common interface. Wifi Direct to handle Wifi Direct tasks. Announce service when using WiFi direct
