+++ p2p/src/main/AndroidManifest.xml
@@ -1,6 +1,9 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-
-    package="sintulabs.p2p">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="sintulabs.p2p">
+    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
+    <uses-permission android:name="android.permission.CHANGE_WIFI_STATE" />
+    <uses-permission android:name="android.permission.CHANGE_NETWORK_STATE" />
+    <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
     <application android:allowBackup="true" android:label="@string/app_name"
         android:supportsRtl="true">
